move the shared enums to the commons subpackage
