+++ hydrolib/core/dflowfm/common/__init__.py
@@ -1,5 +1,17 @@
-"""Common model types shared across D-Flow FM file formats."""
+"""Common D-Flow FM model types shared across subpackages."""
 
-from .models import LocationType, Operand
+from .models import (
+    AveragingType,
+    DataFileType,
+    InterpolationMethod,
+    LocationType,
+    Operand,
+)
 
-__all__ = ["LocationType", "Operand"]
+__all__ = [
+    "LocationType",
+    "Operand",
+    "DataFileType",
+    "InterpolationMethod",
+    "AveragingType",
+]
